+++ src/main/kotlin/net/papierkorb2292/command_crafter/editor/CommandCrafterLanguageClient.kt
@@ -22,6 +22,9 @@ interface CommandCrafterLanguageClient : LanguageClient, EditorClientFileFinder 
     @JsonNotification
     fun logMinecraftMessage(message: ConsoleMessage)
 
+    @JsonNotification
+    fun modVersion(version: String)
+
     @JsonRequest
     fun getFileContent(path: String): CompletableFuture<String>