+++ editor_extensions/vscode/commandcrafter/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - The extension now tries to fill in the "Feature Config" with default values provided by the language server
 - The extension now recognizes if Spyglass is installed and asks to disable semantic tokens for one of two extensions
+- The extension and the mod now notify each other of their version
 
 ## 0.5.0